Separated classes into packages
