+++ src/main/java/Duke.java
@@ -1,3 +1,7 @@
+import duke.*;
+import duke.Command.Command;
+import duke.Exception.DukeException;
+
 public class Duke {
     private TaskList tl;
     private Storage storage;
