--- src/main/java/UnmarkCommand.java
@@ -1,16 +1,0 @@
-public class UnmarkCommand extends Command {
-    private int index;
-
-    public UnmarkCommand(int index) {
-        this.index = index;
-    }
-
-    @Override
-    public void execute(TaskList tasks, UI ui, Storage storage) throws InvalidTaskException {
-        if (index < 1 || index > tasks.items) {
-            throw new InvalidTaskException(index);
-        }
-        ui.showConfirmation(tasks.tasks.get(index - 1).markAsUndone());
-        storage.saveToFile(tasks.tasks);
-    }
-}
