+++ src/main/java/duke/Command/Command.java
@@ -1,3 +1,10 @@
+package duke.Command;
+
+import duke.Exception.DukeException;
+import duke.Storage;
+import duke.TaskList;
+import duke.UI;
+
 public abstract class Command {
 
     public abstract void execute(TaskList tasks, UI ui, Storage storage) throws DukeException;
