+++ src/main/java/duke/Command/DeleteCommand.java
@@ -1,3 +1,10 @@
+package duke.Command;
+
+import duke.Exception.InvalidTaskException;
+import duke.Storage;
+import duke.TaskList;
+import duke.UI;
+
 public class DeleteCommand extends Command {
 
     private final int index;
@@ -8,10 +15,10 @@ public class DeleteCommand extends Command {
 
     @Override
     public void execute(TaskList tasks, UI ui, Storage storage) throws InvalidTaskException {
-        if (index < 1 || index > tasks.items) {
+        if (index < 1 || index > tasks.getItems()) {
             throw new InvalidTaskException(index);
         }
         ui.showConfirmation(tasks.deleteTask(index -1));
-        storage.saveToFile(tasks.tasks);
+        storage.saveToFile(tasks.getTasks());
     }
 }
