+++ src/main/java/duke/Command/ToDoCommand.java
@@ -1,3 +1,10 @@
+package duke.Command;
+
+import duke.Storage;
+import duke.TaskList;
+import duke.Task.ToDo;
+import duke.UI;
+
 public class ToDoCommand extends Command {
 
     private String name;
@@ -10,6 +17,6 @@ public class ToDoCommand extends Command {
     public void execute(TaskList tasks, UI ui, Storage storage) {
         ToDo td = new ToDo(name);
         ui.showConfirmation(tasks.addTask(td));
-        storage.saveToFile(tasks.tasks);
+        storage.saveToFile(tasks.getTasks());
     }
 }
