+++ src/main/java/duke/Exception/DukeException.java
@@ -1,3 +1,5 @@
+package duke.Exception;
+
 public abstract class DukeException extends Exception {
 
     public DukeException(String message) {
