+++ src/main/java/duke/Exception/InvalidInputException.java
@@ -1,3 +1,5 @@
+package duke.Exception;
+
 public class InvalidInputException extends DukeException {
 
     public InvalidInputException(String input) {
