+++ src/main/java/duke/Storage.java
@@ -1,3 +1,8 @@
+package duke;
+
+import duke.Exception.NoSuchFileException;
+import duke.Task.Task;
+
 import java.io.*;
 import java.util.ArrayList;
 import java.util.List;
