+++ src/main/java/duke/Task/Deadline.java
@@ -1,3 +1,5 @@
+package duke.Task;
+
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
