+++ src/main/java/duke/Task/Task.java
@@ -1,3 +1,5 @@
+package duke.Task;
+
 import java.time.LocalDateTime;
 
 public abstract class Task {
